Extracted CoordTransform.
