+++ examples/read_write_ogr.rs
@@ -1,9 +1,10 @@
 use gdal::errors::Result;
-use gdal::spatial_ref::{CoordTransform, SpatialRef};
+use gdal::spatial_ref::SpatialRef;
 use gdal::Dataset;
-use gdal::{vector::*, DriverManager};
+use gdal::{DriverManager, vector::*};
 use std::fs;
 use std::path::Path;
+use gdal::spatial_ref::CoordTransform;
 
 fn run() -> Result<()> {
     let dataset_a = Dataset::open(Path::new("fixtures/roads.geojson"))?;
