+++ src/spatial_ref/mod.rs
@@ -3,7 +3,9 @@
 //! <https://gdal.org/api/ogr_srs_api.html>
 
 mod srs;
+mod transform;
 mod transform_opts;
 
-pub use srs::{AxisOrientationType, CoordTransform, SpatialRef};
+pub use srs::{AxisOrientationType, SpatialRef};
+pub use transform::CoordTransform;
 pub use transform_opts::CoordTransformOptions;
